requirements.txt
Pillow
numpy
paramiko
